Refactoring RBM
